Fix the scale information loss bug when export quantize model

+++ paddle2onnx/mapper/onnx_helper.h
@@ -530,12 +530,12 @@ bool OnnxHelper::TryGetTensorValue(const std::string& name,
             value->assign(val.begin(), val.end());
             return true;
           } else if (dtype == ONNX_NAMESPACE::TensorProto::FLOAT) {
-            std::vector<int32_t> val(nums, 0);
+            std::vector<float> val(nums, 0);
             memcpy(val.data(), tensor->raw_data().data(), nums * sizeof(float));
             value->assign(val.begin(), val.end());
             return true;
           } else if (dtype == ONNX_NAMESPACE::TensorProto::DOUBLE) {
-            std::vector<int32_t> val(nums, 0);
+            std::vector<double> val(nums, 0);
             memcpy(val.data(), tensor->raw_data().data(),
                    nums * sizeof(double));
             value->assign(val.begin(), val.end());
